PB-1855 Group per incident, add printing incident
This avoids having too many entries in the table of contents of the status page on the right side.

The preview of the status page on the start page was adapted to show that everything is ok. I also added a comment on the possible values to make it easy to edit this when people are in a hurry.

To unify structuring, the titles were adapted to start with the data in square brackets plus a short title.

The printing incident was added from an export of the history of the CMS. The incident announcement is not available online anymore on geo.admin.ch.