PB-1855 Remove page "SPARQL Service"
It's not a `*geo.admin.ch` service, so we decided to remove it from the Tech Docs. There is still a page in CMS about it: "[Linked Data Service: linking geodata](https://www.geo.admin.ch/en/linked-data-service-linking-geodata/)".

I slightly reworded the summary on category "Access Data" which previously contained the SPARQL service.