PB-1855 Remove page "Atom Feed"
This is a service we plan to shut down soon, so no use to keep it for the new technical documentation.

I slightly reworded the summary on category "Download data". Now, the STAC API is the only entry there, so we might want to consider moving that elsewhere. That is out of scope here though.