PB-1798 Align section titles
The other section titles were self-explaining ("Feature Selection", "Parameter Overview") while others had the name of the parameter of the title. I am not strictly in favor of one over the other but the mixture makes it confusing. So I aligned to the descriptive version for all (e.g. "layers" --> "Layer Parameters").