PB-1798 Rework content of page JS API
I tried to reduce the overly verbose text to the essential. This is going to be decommissioned soon, so we don't want to put any additional effort into it.

We might want to include the examples from the [API Examples page](https://api3.geo.admin.ch/api/examples.html). It's not clear at that moment what to keep from there, so that is for later.

It might make sense to keep the link to the page https://geoadmin.github.io/ol3/apidoc/ but I am not sure what that is. Is it the detailed specs for the JS API?