PB-1798 Add section on how to use swisssearch
In the [Parameter List](https://backend.geo.admin.ch/fileservice/sdweb-docs-prod-geoadminch-files/files/2024/12/04/a7d324a1-98dc-4b9e-88b0-a5dd87d5317e.pdf) linked on [the corresponding CMS help page](https://www.geo.admin.ch/en/map-viewer-functions-share), there is a short mentioning of how to use `swisssearch_autoselect` in an iframe:

> Hier auch ein iFrame Beispiel
<iframe src="https://map.geo.admin.ch/#/embed?layers=ch.bfs.gebaeude_wohnungs_register&swisssearch=Holzikofenweg 36, 3007 Bern&swisssearch_autoselect=true" name="IframeGeoadmin"></iframe>

That was still missing on this page. It is briefly mentioned on page "URL Parameters" under parameter `swissearch` but it's better to have an example to look at. In the description of  parameter `swisssearch` there is already a link to the iframe page. Ideally, it would be even an anchor to the corresponding subsection ("Center on a search result") but I thought that this is too fragile. The section title could change and then the link would be broken.