PB-1939 Add section about INTERLIS / ÖREB
The content is adapted from api3 page "[OEREB/RDPPF: Feature Service](https://api3.geo.admin.ch/services/oerebservices.html)".

I did not put the examples into an `ExampleCodeBlock` because

1. The query stays clickable such that the result can be checked conveniently within the browser.
2. The responses are very long.

I put the section to the end of the text because it makes it clear that this is an edge case, not a core functionality.