PB-1939 Rewrite the section on parameter layerDefs
We learnt that the "queryable" layers are still a thing, so we list them here as they are reported in section "[Which layers are queryable?](https://api3.geo.admin.ch/api/faq/index.html#which-layers-are-queryable)" in the api3 FAQ. That list is populated from the result of an API endpoint. However, for simplicity, we hard code it here as these layers are unlikely to change soon.

Also, I added an introduction sentence and tried to improve the structure of the format explanation.