PB-1877 Directly link to Terms of Use
The page currently pointed at is going to change with the go-live of the Tech Docs.

This is labelled with PB-1877 because it's a small fix that I want to use to test whether CodeBuild etc works as expected. Ultimately, the PR out of this change should contain a description with a preview link to a version on DEV of the Tech Docs that corresponds to the status of the branch.