PB-1877 Set base path for test links
For test links, we need to take into account that there is a "/preview/feat-PB-1234-my-branch" in between. If we don't add this to the static asset links, it will result in broken styling, missing images and broken page links.

See ["Base URL"](https://vitepress.dev/guide/asset-handling#base-url) in the VitePress docs.