PB-1969 Status page: Add templates
This is to make it easier for the person working on the incident repsonse to choose the right wording.

I also changed the danger icon from a red cross (❌) to an alarm light (🚨). A red cross is rather use for a build error while the police light better reflects the urgency of the situation.