PB-1969 Make diagram more readable
Increase the font sizes and make the diagram more compact such that it appears bigger in the rendered final version of the web page.

Wrapping the image into a link that opens a new tab does not work because then the image is not going through Vitpress' asset pipeline, so the image is not found in the PROD build. We could put the image into the "public" folder (where also the icons etc are) but that goes against the advice to keep this folder's content minimal. Also it might go against the idea to have versioned assets to simplify deployment later.

Add a JS library for an image viewer with zooming capabilities just for this image seems overkill.