PB-2051 Improve language on startpage #patch
Just a minor fix in language to the Release Notes section on the homepage. "The latest releases of software" sounds awkward and "news on changes" can be shortened.

The main intent of this change is to be able to test the deploy to INT and PROD after the changes in https://github.com/geoadmin/doc-tech/pull/68 (PB-2051).