PB-2084 Fix table
The table was broken because the formatter interpreted the `|` in the URL as a Markdown column separator.

The `layers` parameter used to separate layers with the `|` symbol; today it's with a semicolon (`;`). The new URL takes this into account.

I did not understand why there was  a `layers=WMS|http://wms.geo.admin.ch/?` in the URL, so I removed it.